spt update

+++ database/migrations/2023_08_11_024157_create_tx_neraca.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->increments('id');
             $table->integer('no_akun');
             $table->string('nama_akun');
-            $table->integer('saldo');
+            $table->float('saldo');
             $table->string('attribute1')->nullable();
             $table->string('attribute2')->nullable();
             $table->string('attribute3')->nullable();
